refactor(ui): decompose game-carousel into modular directory structure
Convert 914-line game-carousel.tsx into a well-organized directory:
- types.ts: CarouselGame, GameCarouselProps, CarouselSortValue, CarouselSortOption
- constants.ts: CAROUSEL_SORT_OPTIONS, ANIMATION_DURATION, SWIPE_THRESHOLD
- math.ts: CardPosition interface + calculateCardPositions utility
- hooks/useSwipe.ts, hooks/useKeyboardNavigation.ts
- components/NavButton, DotsIndicator, AutoPlayButton, SortDropdown
- GameCarousel.tsx: main component (React.memo)
- GameCarouselSkeleton.tsx: loading skeleton
- index.ts: barrel exports preserving existing public API

All consumers import from '@/components/ui/data-display/game-carousel'
which now resolves to the directory's index.ts — zero import changes needed.
