+++ apps/web/src/components/ui/data-display/game-carousel/components/index.ts
@@ -1,0 +1,4 @@
+export { NavButton } from './NavButton';
+export { DotsIndicator } from './DotsIndicator';
+export { AutoPlayButton } from './AutoPlayButton';
+export { SortDropdown } from './SortDropdown';
