+++ apps/web/src/components/ui/data-display/game-carousel/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useSwipe } from './useSwipe';
+export { useKeyboardNavigation } from './useKeyboardNavigation';
