+++ apps/web/src/components/ui/data-display/game-carousel/hooks/useKeyboardNavigation.ts
@@ -1,0 +1,25 @@
+'use client';
+
+import { useEffect } from 'react';
+
+/**
+ * Custom hook for keyboard navigation
+ */
+export function useKeyboardNavigation(onPrev: () => void, onNext: () => void, enabled: boolean) {
+  useEffect(() => {
+    if (!enabled) return;
+
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'ArrowLeft') {
+        e.preventDefault();
+        onPrev();
+      } else if (e.key === 'ArrowRight') {
+        e.preventDefault();
+        onNext();
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+    return () => window.removeEventListener('keydown', handleKeyDown);
+  }, [onPrev, onNext, enabled]);
+}
