+++ apps/web/src/components/ui/data-display/game-carousel/hooks/useSwipe.ts
@@ -1,0 +1,34 @@
+'use client';
+
+import React, { useCallback, useRef } from 'react';
+
+import { SWIPE_THRESHOLD } from '../constants';
+
+/**
+ * Custom hook for touch/swipe handling
+ */
+export function useSwipe(onSwipeLeft: () => void, onSwipeRight: () => void) {
+  const touchStartX = useRef<number>(0);
+  const touchEndX = useRef<number>(0);
+
+  const handleTouchStart = useCallback((e: React.TouchEvent) => {
+    touchStartX.current = e.touches[0].clientX;
+  }, []);
+
+  const handleTouchMove = useCallback((e: React.TouchEvent) => {
+    touchEndX.current = e.touches[0].clientX;
+  }, []);
+
+  const handleTouchEnd = useCallback(() => {
+    const diff = touchStartX.current - touchEndX.current;
+    if (Math.abs(diff) > SWIPE_THRESHOLD) {
+      if (diff > 0) {
+        onSwipeLeft();
+      } else {
+        onSwipeRight();
+      }
+    }
+  }, [onSwipeLeft, onSwipeRight]);
+
+  return { handleTouchStart, handleTouchMove, handleTouchEnd };
+}
