+++ apps/web/src/components/ui/data-display/game-carousel/index.ts
@@ -1,0 +1,9 @@
+export { GameCarousel } from './GameCarousel';
+export { GameCarouselSkeleton } from './GameCarouselSkeleton';
+export type {
+  GameCarouselProps,
+  CarouselGame,
+  CarouselSortValue,
+  CarouselSortOption,
+} from './types';
+export { CAROUSEL_SORT_OPTIONS } from './constants';
